Corrected typo/spelling-error
I wasn't really paying much attention to the whole/entire document. Wanted to leave a note somewhere, but figured it was better to just submit this instead, and you can correct it at your leisure. There's a typo in the text, where you missed a keystroke when spelling the work `submitting`. I will try to look thru the document again, and see if i spot anything new or different. But I've already re-read it like two or three times now (maybe more), while going thru the how-to, and i'm honestly a little too lazy to keep at it. 

I've also opened an issue, for adjustments to be made to readme. Though, i've lost my train of thought on how best to go about it. Maybe the two can be dealt with in one stroke. (two birds, one stone...)